Removed NullSdkPathException. Relocated library exceptions.
