+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/appengine/sdk/CloudSdkPanel.java
@@ -17,10 +17,9 @@
 package com.google.cloud.tools.intellij.appengine.sdk;
 
 import com.google.api.client.repackaged.com.google.common.annotations.VisibleForTesting;
-import com.google.cloud.tools.appengine.api.exceptions.AppEngineException;
-import com.google.cloud.tools.appengine.api.exceptions.BadSdkLocationException;
-import com.google.cloud.tools.appengine.api.exceptions.NotAFileException;
-import com.google.cloud.tools.appengine.api.exceptions.NullSdkPathException;
+import com.google.cloud.tools.appengine.api.AppEngineException;
+import com.google.cloud.tools.appengine.api.BadSdkLocationException;
+import com.google.cloud.tools.appengine.api.NotAFileException;
 import com.google.cloud.tools.appengine.cloudsdk.CloudSdk;
 import com.google.cloud.tools.intellij.util.GctBundle;
 
@@ -92,9 +91,7 @@ public class CloudSdkPanel {
     } catch (AppEngineException aee) {
       cloudSdkDirectoryField.getTextField().setForeground(JBColor.red);
       warningMessage.setVisible(true);
-      if (aee instanceof NullSdkPathException) {
-        warningMessage.setText(GctBundle.message("appengine.cloudsdk.location.missing.message"));
-      } else if (aee instanceof BadSdkLocationException) {
+      if (aee instanceof BadSdkLocationException) {
         warningMessage.setText(GctBundle.message("appengine.cloudsdk.location.directory.invalid"));
       } else if (aee instanceof NotAFileException) {
         warningMessage.setText(GctBundle.message("appengine.cloudsdk.location.invalid.message"));
